Added the following options to TagOptionSingleton: id3v23DefaultTextEncoding:This is the default text encoding to use for new v23 frames, when unicode is required UTF16 will always be used because that is the only valid option for v23.
id3v24DefaultTextEncoding:This is the default text encoding to use for new v24 frames, it defaults to simple ISO8859 but by changing this value you could always used UTF8 for example whether you needed to or not

id3v24UnicodeTextEncoding:This is text encoding to use for new v24 frames when unicode is required, it defaults to UTF16 just because this encoding is understand by all ID3 versions

resetTextEncodingForExistingFrames:When writing frames if this is set to true then the frame will be written using the defaults disregarding the text encoding originally used to createthe frame.

also added tests, and removed some options from TagOptionSingleton because not implemented and probably never will be.


git-svn-id: https://svn.java.net/svn/jaudiotagger~svn/trunk@301 922cec90-98bc-ecba-b0a9-867c771b4a74